Add ability to search specific passwords generated
Using the findall function from re to find a specific portion of the password.txt file that has the password associated with specified account type and username. Using the pattern variable and format function together to get the desired pattern. Also have a message displayed in the event that it cannot find said account type and username.