Made some changes to README.md
Changed Function section to Option to since each option are no longer functions. Added step-by-step explantion of how "Clear All Data", "Display Info", and "Specific Search" Works. Moved "BE SURE TO UNZIP FILE OR THE PASSWOD.TXT WILL NOT GENERATE" to the bottom under "Potential Issues" section and state what issue it was meant to solve. Added issue "Two entries on the same line" with "FIND WHERE thE END OF THE FIRST ENTRY ENDS AND HIT THE ENTER KEY" as the solution to the "Potential Issues" section